LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/registered-property-initial-expected.txt


FAIL Initial values of registered properties can be referenced when no custom properties are explicitly set. assert_equals: expected "25px" but got "calc(10px + 15px)"
